# lesson-14/project-backend/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "My movies app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of my movies app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://my-movies.com/api/v1"
    }
  ],
  "tags": [
    {
      "name": "Movies",
      "description": "Operations about movies."
    }
  ],
  "paths": {
    "/movies": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Get list of movies",
        "operationId": "getMovies",
        "description": "Get list of movies with this endpoint",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "example": "title"
            },
            "description": "All movies's fields can be mentioned. Default value - _id"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "example": "asc"
            }
          },
          {
            "in": "query",
            "name": "minReleaseYear",
            "schema": {
              "type": "number",
              "example": 1885
            }
          },
          {
            "in": "query",
            "name": "maxReleaseYear",
            "schema": {
              "type": "number",
              "example": 2030
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found movies!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found movies!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "total",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/movie"
                          }
                        },
                        "page": {
                          "type": "number",
                          "example": 2
                        },
                        "perPage": {
                          "type": "number",
                          "example": 4
                        },
                        "total": {
                          "type": "number",
                          "example": 4
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 4
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": false
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "movie": {
        "type": "object",
        "required": [
          "_id",
          "title",
          "director",
          "releaseYear"
        ],
        "properties": {
          "_id": {
            "description": "Movie's id",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "title": {
            "description": "Movie's title",
            "type": "string",
            "example": "Avatar"
          },
          "director": {
            "description": "Movie's director",
            "type": "string",
            "example": "James Cameron"
          },
          "releaseYear": {
            "description": "Movie release year",
            "type": "number",
            "example": 1885
          },
          "type": {
            "description": "`Movie's type",
            "type": "string",
            "enum": [
              "film",
              "serial"
            ]
          },
          "favorite": {
            "description": "Is movie favorite",
            "type": "boolean",
            "example": false
          },
          "userId": {
            "description": "Movie's parent id",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "poster": {
            "description": "Link to movie's poster",
            "type": "string",
            "example": "https://res.cloudinary.com/uqwgdu/image/upload/image.png"
          }
        }
      }
    },
    "responses": {
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}